waydroid: Optionally start CCodec service from /vendor_extra
On the Pixel 3a with Ubuntu Touch CCodec is used on the host
as determined by debug.stagefright.ccodec=4.

This has the weird effect of starting the missing services
for multimedia playback, but using OMX in the end since.. why not?

For completeness sake, and since on the host both OMX and C2 are
running anyway, just start C2 inside of the WayDroid container as well.